Rename module `content-store` -> `file-store`.
It's not just about content anymore.
