+++ local-modules/doc-server/package.json
@@ -7,8 +7,8 @@
     "weak": "^1.0.1",
 
     "codec": "local",
-    "content-store": "local",
     "doc-common": "local",
+    "file-store": "local",
     "hooks-server": "local",
     "see-all": "local",
     "server-env": "local",
