+++ local-modules/file-store-local/Transactor.js
@@ -2,7 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { Errors } from 'content-store';
+import { Errors } from 'file-store';
 import { CommonBase, InfoError } from 'util-common';
 
 /**
@@ -32,7 +32,7 @@ export default class Transactor extends CommonBase {
      * {object} "Friend" access to the `LocalFile` that spawned this instance.
      * Properties and methods on this are defined in an ad-hoc manner intended
      * to provide just enough access for this class to do its work. See
-     * {@link content-store.FileOp} where the friend is constructed for
+     * {@link file-store.FileOp} where the friend is constructed for
      * documentation on its makeup.
      */
     this._fileFriend = fileFriend;
