+++ local-modules/file-store/main.js
