+++ local-modules/file-store/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "content-store",
+  "name": "file-store",
   "version": "1.0.0",
   "main": "main.js",
